chore: fix measurements
