+++ collections/IngredientGroup.ts
@@ -1,5 +1,12 @@
 import type { CollectionConfig } from "payload";
 
+const measurementMapping: Record<string, string> = {
+	weight: "Weight - grams",
+	volume: "Volume - milliliters",
+	pieces: "Pieces",
+	"drink-volume": "Drink volume - milliliters",
+};
+
 export const IngredientGroups: CollectionConfig = {
 	slug: "ingredient-groups",
 	admin: {
@@ -22,6 +29,22 @@ export const IngredientGroups: CollectionConfig = {
 					type: "relationship",
 					relationTo: "ingredients",
 					required: true,
+					label: "Ingredient",
+				},
+				{
+					name: "measurement",
+					type: "select",
+					required: true,
+					label: "Measurement",
+					options: [
+						{ label: measurementMapping.weight, value: "weight" },
+						{ label: measurementMapping.volume, value: "volume" },
+						{ label: measurementMapping.pieces, value: "pieces" },
+						{
+							label: measurementMapping["drink-volume"],
+							value: "drink-volume",
+						},
+					],
 				},
 				{
 					name: "amount",
